refactor(discord): split message process tests
